Extracted all methods out
