+++ Properties/launchSettings.json
@@ -8,6 +8,10 @@
       "commandName": "Project",
       "commandLineArgs": "CompileDatabase \r\n-p \"/Users/Neskol/MaiAnalysis/A000/\"\r\n-c \"/Users/Neskol/MaiAnalysis/Image/Texture2D/\"\r\n-m \"/Users/Neskol/MaiAnalysis/Sound/\" \r\n-o \"/Users/Neskol/MaiAnalysis/Output_NoBGA/\" "
     },
+    "NoBGA Run by Cabinet": {
+      "commandName": "Project",
+      "commandLineArgs": "CompileDatabase \r\n-g \"2\" \r\n-p \"/Users/Neskol/MaiAnalysis/A000/\"\r\n-c \"/Users/Neskol/MaiAnalysis/Image/Texture2D/\"\r\n-m \"/Users/Neskol/MaiAnalysis/Sound/\" \r\n-o \"/Users/Neskol/MaiAnalysis/Output_NoBGA_ByCab/\" "
+    },
     "With BGA": {
       "commandName": "Project",
       "commandLineArgs": "CompileDatabase \r\n-p \"/Users/Neskol/MaiAnalysis/A000/\"\r\n-c \"/Users/Neskol/MaiAnalysis/Image/Texture2D/\"\r\n-m \"/Users/Neskol/MaiAnalysis/Sound/\"\r\n-v \"/Users/Neskol/MaiAnalysis/DXBGA_HEVC/\" \r\n-o \"/Users/Neskol/MaiAnalysis/Output_HEVC/\" "
